merge newest code 20200411

+++ lib/ui/lerp.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart = 2.6
 part of dart.ui;
 
 /// Linearly interpolate between two numbers.
@@ -10,5 +11,5 @@ double lerpDouble(num a, num b, double t) {
     return null;
   a ??= 0.0;
   b ??= 0.0;
-  return a + (b - a) * t;
+  return a + (b - a) * t as double;
 }
